more thorough refactoring as Dimitry suggested
