+++ freeplane/src/main/java/org/freeplane/view/swing/map/overview/MapOverviewAttachPoint.java
@@ -1,0 +1,5 @@
+package org.freeplane.view.swing.map.overview;
+
+public enum MapOverviewAttachPoint {
+    NORTH_WEST, NORTH_EAST, SOUTH_WEST, SOUTH_EAST
+}
